feat: refactor week types
